requirements.txt
python-dotenv
boto3
pytest
pylint
psycopg2-binary
confluent_kafka
ipykernal